Refactor file locking and add standby middleware
- Renamed `lock` and `unlock` methods in `FileLock` class to `acquire_lock` and `release_lock` for clarity.
- Introduced `StandbyMiddleware` to manage request handling during standby mode, allowing only the `/ready` endpoint.
- Moved middleware-related functions to a new `middleware.py` file for better organization.
- Updated the main application to include the new middleware and ensure proper server initialization and shutdown logging.
